+++ docs/components/homepage/steps/step01.ts
@@ -20,7 +20,7 @@ export default {
     {
       label: 'Full Entity API',
       href: '/docs/ref_entity',
-    }
+    },
   ],
   files: [
     {
@@ -33,7 +33,7 @@ export default {
   allowApiCrud: true,
 })
 export class Task {
-  @Fields.cuid()
+  @Fields.id()
   id!: string
 
   @Fields.string()