+++ docs/components/homepage/steps/step05.ts
@@ -15,7 +15,7 @@ export default {
   allowApiCrud: true,
 })
 export class Task {
-  @Fields.cuid()
+  @Fields.id()
   id!: string
 
   @Fields.string({