+++ docs/interactive/src/content/tutorial/1-basics/4-validation/3-custom-validations/_files/shared/Task.ts
@@ -4,7 +4,7 @@ import { Entity, Fields, Validators } from 'remult'
   allowApiCrud: true,
 })
 export class Task {
-  @Fields.uuid()
+  @Fields.id()
   id = ''
 
   @Fields.string()