+++ docs/interactive/src/content/tutorial/1-basics/7-auth/1-introduction/_files/shared/Task.ts
@@ -4,7 +4,7 @@ import { Entity, Fields, remult } from 'remult'
   allowApiCrud: true, // <-- Change this line
 })
 export class Task {
-  @Fields.uuid()
+  @Fields.id()
   id = ''
 
   @Fields.string({