+++ docs/interactive/src/content/tutorial/4-concepts/7-access-control/1-field-level-control/_files/shared/Task.ts
@@ -4,7 +4,7 @@ import { Entity, Fields, remult, getEntityRef } from 'remult'
   allowApiCrud: true,
 })
 export class Task {
-  @Fields.uuid()
+  @Fields.id()
   id = ''
 
   @Fields.string({