+++ docs/interactive/src/content/tutorial/4-concepts/7-access-control/3-filtering-rows-based-on-User-Permissions/_files/shared/Task.ts
@@ -9,7 +9,7 @@ import { Entity, Fields, remult, ForbiddenError } from 'remult'
   allowApiCrud: true,
 })
 export class Task {
-  @Fields.uuid()
+  @Fields.id()
   id = ''
 
   @Fields.string({