+++ docs/interactive/src/content/tutorial/4-concepts/7-access-control/5-sql-relations-filter-for-user-permissions/_files/shared/Task.ts
@@ -5,7 +5,7 @@ import { Entity, Fields, remult, ForbiddenError, Filter } from 'remult'
   allowApiCrud: true,
 })
 export class Task {
-  @Fields.uuid()
+  @Fields.id()
   id = ''
 
   @Fields.string({ required: true })