+++ examples/nuxt-todo/shared/task.ts
@@ -6,7 +6,7 @@ import { Allow, Entity, Fields, Validators } from 'remult'
   allowApiDelete: 'admin',
 })
 export class Task {
-  @Fields.cuid()
+  @Fields.id()
   id = ''
 
   @Fields.string<Task>({