+++ projects/core/src/remult3/Fields.ts
@@ -1,5 +1,3 @@
-import { createId } from '@paralleldrive/cuid2'
-import { v4 as uuid } from 'uuid'
 import type { ClassType } from '../../classType.js'
 import type { FieldOptions, ValueConverter } from '../column-interfaces.js'
 import type { Remult } from '../context.js'
@@ -172,61 +170,73 @@ export class Fields {
     )
   }
 
+  /** @deprecated use `@Fields.id()` instead */
   static uuid<entityType = unknown>(
     ...options: (
       | FieldOptions<entityType, string>
       | ((options: FieldOptions<entityType, string>, remult: Remult) => void)
     )[]
   ): ClassFieldDecorator<entityType, string | undefined> {
-    return Field(
-      () => String as any,
-      {
-        allowApiUpdate: false,
-        defaultValue: () => uuid(),
-        saving: (_, r) => {
-          if (!r.value) r.value = uuid()
-        },
-      },
+    return Fields.id({
+      idFactory: () => crypto.randomUUID(),
       ...options,
-    )
+    })
   }
+
+  private static _defaultIdFactory: () => string = () => crypto.randomUUID()
+
+  static get defaultIdFactory(): () => string {
+    return this._defaultIdFactory
+  }
+
+  static set defaultIdFactory(value: () => string) {
+    this._defaultIdFactory = value
+    remultStatic.defaultIdFactory = value
+  }
+
   /**
-   * A CUID (Collision Resistant Unique Identifier) field.
-   * This id value is determined on the backend on insert, and can't be updated through the API.
-   * The CUID is generated using the `@paralleldrive/cuid2` npm package.
+   * Defines a field that will be used as the id of the entity.
+   * By default it will use `crypto.randomUUID` to generate the id.
+   *
+   * You can change the algorithm used to generate the id by setting the `Fields.defaultIdFactory`
+   * to a different function like:
+   *
+   * This needs to be done in a shared file to be accessible frontend and backend.
+   *
+   * ```ts
+   * import { createId } from '@paralleldrive/cuid2'
+   * Fields.defaultIdFactory = () => createId()
+   * ```
+   *
+   * You can also pass an id factory as an option to the `@Fields.id` to have a different value locally.
+   * @example
+   * ```ts
+   * import { createId } from '@paralleldrive/cuid2'
+   * // import { v4 as uuid } from 'uuid'
+   *
+   * class MyEntity {
+   *   \@Fields.id({
+   *     idFactory: () => createId()
+   *     // idFactory: () => uuid()
+   *   })
+   *   id: string = '';
+   * }
+   * ```
    */
-
-  static cuid<entityType = unknown>(
-    ...options: (
-      | FieldOptions<entityType, string>
-      | ((options: FieldOptions<entityType, string>, remult: Remult) => void)
-    )[]
+  static id<entityType = unknown>(
+    options?: FieldOptions<entityType, string> & { idFactory?: () => string },
   ): ClassFieldDecorator<entityType, string | undefined> {
-    return Field(
-      () => String as any,
-      {
-        allowApiUpdate: false,
-        defaultValue: () => createId(),
-        saving: (_, r) => {
-          if (!r.value) r.value = createId()
-        },
+    let idFactory = options?.idFactory ?? remultStatic.defaultIdFactory
+
+    return Field(() => String as any, {
+      allowApiUpdate: false,
+      defaultValue: () => (options?.allowNull ? null! : idFactory()),
+      saving: (_, r) => {
+        if (!r.value && !options?.allowNull) r.value = idFactory()
       },
       ...options,
-    )
+    })
   }
-
-  // static id<entityType = unknown>(
-  //   options?: FieldOptions<entityType, string> & { idFactory?: () => string },
-  // ): ClassFieldDecorator<entityType, string | undefined> {
-  //   let buildId = options?.idFactory ?? createId
-  //   return Field(() => String, {
-  //     allowApiUpdate: false,
-  //     defaultValue: () => buildId(),
-  //     saving: (_, r) => {
-  //       if (!r.value) r.value = buildId()
-  //     },
-  //   })
-  // }
 
   /**
  * Defines a field that can hold a value from a specified set of string literals.
@@ -337,6 +347,10 @@ export class Fields {
     return Field(() => Boolean as any, ...options)
   }
 }
+
+// Initialize remultStatic.defaultIdFactory with the default value
+remultStatic.defaultIdFactory = Fields.defaultIdFactory
+
 export class Relations {
   /**
    * Define a to-one relation between entities, indicating a one-to-one relationship.