+++ projects/core/src/remult3/IdEntity.ts
@@ -2,6 +2,6 @@ import { Fields } from './Fields.js'
 import { EntityBase } from './RepositoryImplementation.js'
 
 export class IdEntity extends EntityBase {
-  @Fields.uuid()
+  @Fields.id()
   id!: string
 }