+++ projects/create-remult/templates/auth/auth.js/shared/src/demo/auth/User.ts
@@ -17,7 +17,7 @@ import bcrypt from "bcryptjs";
   },
 })
 export class User {
-  @Fields.cuid()
+  @Fields.id()
   id = "";
 
   @Fields.string({ required: true, validate: Validators.unique() }) // User's name, required field and must be unique