+++ projects/test-servers/sveltekit-server/src/shared/Task.ts
@@ -15,7 +15,7 @@ import {
   },
 })
 export class Task {
-  @Fields.uuid()
+  @Fields.id()
   id!: string
 
   @Fields.string((options, remult) => {